refactor(prompts): split out prompts into package
