+++ packages/@ionic/cli-framework-prompts/README.md
@@ -1,0 +1,1 @@
+# @ionic/cli-framework-prompts
