+++ packages/@ionic/cli-framework-prompts/lint-staged.config.js
@@ -1,0 +1,1 @@
+module.exports = require('../../../lint-staged.config.base');
