+++ packages/@ionic/cli-framework-prompts/package.json
@@ -1,0 +1,44 @@
+{
+  "name": "@ionic/cli-framework-prompts",
+  "version": "1.0.0",
+  "description": "The interactive prompts portion of Ionic CLI Framework",
+  "homepage": "https://ionicframework.com/",
+  "author": "Ionic Team <hi@ionic.io> (https://ionicframework.com)",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "engines": {
+    "node": ">=8.9.4"
+  },
+  "files": [
+    "dist/",
+    "LICENSE",
+    "README.md"
+  ],
+  "scripts": {
+    "clean": "rimraf dist",
+    "lint": "tslint --project tsconfig.json",
+    "build": "npm run clean && tsc",
+    "watch": "tsc -w --preserveWatchOutput",
+    "test": "jest --maxWorkers=4",
+    "prepublishOnly": "npm run build"
+  },
+  "license": "MIT",
+  "dependencies": {
+    "@ionic/utils-terminal": "1.0.5",
+    "debug": "^4.0.0",
+    "inquirer": "^7.0.0",
+    "tslib": "^1.9.0"
+  },
+  "devDependencies": {
+    "@types/debug": "^4.1.1",
+    "@types/inquirer": "0.0.43",
+    "@types/jest": "^24.0.3",
+    "@types/node": "~8.10.54",
+    "jest": "^23.0.1",
+    "jest-cli": "^24.1.0",
+    "lint-staged": "^9.2.0",
+    "ts-jest": "^23.10.1",
+    "tslint": "^5.9.1",
+    "typescript": "~3.6.3"
+  }
+}
