+++ packages/@ionic/cli-framework-prompts/tslint.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../../tslint.base.json"
+}
