+++ packages/@ionic/cli-framework/package.json
@@ -19,6 +19,7 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@ionic/cli-framework-prompts": "1.0.0",
     "@ionic/utils-array": "1.2.1",
     "@ionic/utils-fs": "2.0.7",
     "@ionic/utils-object": "1.0.5",
@@ -28,7 +29,6 @@
     "@ionic/utils-terminal": "1.0.5",
     "chalk": "^2.3.0",
     "debug": "^4.0.0",
-    "inquirer": "^7.0.0",
     "lodash": "^4.17.5",
     "log-update": "^3.2.0",
     "minimist": "^1.2.0",
@@ -43,7 +43,6 @@
   },
   "devDependencies": {
     "@types/debug": "^4.1.1",
-    "@types/inquirer": "0.0.43",
     "@types/jest": "^24.0.3",
     "@types/lodash": "^4.14.104",
     "@types/minimist": "^1.2.0",
