+++ packages/ionic/src/lib/index.ts
@@ -1,4 +1,5 @@
-import { LOGGER_LEVELS, createPromptModule } from '@ionic/cli-framework';
+import { LOGGER_LEVELS } from '@ionic/cli-framework';
+import { createPromptModule } from '@ionic/cli-framework-prompts';
 import { prettyPath } from '@ionic/cli-framework/utils/format';
 import { TERMINAL_INFO } from '@ionic/utils-terminal';
 import * as Debug from 'debug';
